+++ crates/canisters/root/src/lib.rs
@@ -1,8 +1,4 @@
-use icu::{
-    interface::state::create_canisters,
-    prelude::*,
-    state::{CanisterAttributes, CanisterRegistry},
-};
+use icu::{interface::state::create_canisters, prelude::*};
 
 //
 // ICU
@@ -13,23 +9,26 @@ icu_start_root!("root");
 #[update]
 async fn init_async() {
     register_canisters();
+    register_canister();
     create_canisters(&[]).await.unwrap();
 }
 
+fn register_canister() {}
+
 // register_canisters
 fn register_canisters() {
-    let canisters: &[(&'static str, CanisterAttributes, &'static [u8])] = &[(
-        "test",
-        CanisterAttributes {
-            auto_create: true,
-            is_sharded: false,
-        },
-        include_bytes!("../../../../.dfx/local/canisters/test/test.wasm.gz"),
-    )];
+    // let canisters: &[(&'static str, CanisterAttributes, &'static [u8])] = &[(
+    //     "test",
+    //     CanisterAttributes {
+    //         auto_create: true,
+    //         is_sharded: false,
+    //     },
+    //     include_bytes!("../../../../.dfx/local/canisters/test/test.wasm.gz"),
+    // )];
 
-    for (path, def, wasm) in canisters {
-        CanisterRegistry::add_canister(path, def, wasm).unwrap();
-    }
+    // for (path, def, wasm) in canisters {
+    //     CanisterRegistry::insert(path, def, wasm).unwrap();
+    // }
 }
 
 export_candid!();