+++ crates/icu/src/memory/app/app_state.rs
@@ -1,6 +1,8 @@
 use crate::{
-    Log, ic::structures::Cell, icu_register_memory, impl_storable_unbounded, log,
-    memory::APP_STATE_MEMORY_ID,
+    Error, Log,
+    ic::structures::Cell,
+    icu_register_memory, impl_storable_unbounded, log,
+    memory::{APP_STATE_MEMORY_ID, MemoryError},
 };
 use candid::CandidType;
 use derive_more::Display;
@@ -36,6 +38,10 @@ pub enum AppStateError {
 pub struct AppState {}
 
 impl AppState {
+    //
+    // INTERNAL ACCESSORS
+    //
+
     pub fn with<R>(f: impl FnOnce(&Cell<AppStateData>) -> R) -> R {
         APP_STATE.with_borrow(|s| f(s))
     }
@@ -44,15 +50,9 @@ impl AppState {
         APP_STATE.with_borrow_mut(|s| f(s))
     }
 
-    #[must_use]
-    pub fn get_data() -> AppStateData {
-        Self::with(Cell::get)
-    }
-
-    // set_data
-    pub fn set_data(data: AppStateData) {
-        Self::with_mut(|cell| cell.set(data));
-    }
+    //
+    // METHODS
+    //
 
     #[must_use]
     pub fn get_mode() -> AppMode {
@@ -68,8 +68,7 @@ impl AppState {
         });
     }
 
-    // command
-    pub fn command(cmd: AppCommand) -> Result<(), AppStateError> {
+    pub fn command(cmd: AppCommand) -> Result<(), Error> {
         let old_mode = Self::with(|cell| cell.get().mode);
 
         let new_mode = match cmd {
@@ -79,7 +78,7 @@ impl AppState {
         };
 
         if old_mode == new_mode {
-            return Err(AppStateError::AlreadyInMode(old_mode));
+            return Err(MemoryError::from(AppStateError::AlreadyInMode(old_mode)))?;
         }
 
         Self::set_mode(new_mode);
@@ -87,6 +86,19 @@ impl AppState {
         log!(Log::Ok, "app: mode changed {old_mode} -> {new_mode}");
 
         Ok(())
+    }
+
+    //
+    // IMPORT & EXPORT
+    //
+
+    pub fn import(data: AppStateData) {
+        Self::with_mut(|cell| cell.set(data));
+    }
+
+    #[must_use]
+    pub fn export() -> AppStateData {
+        Self::with(Cell::get)
     }
 }