+++ crates/icu/src/memory/canister/child_index.rs
@@ -1,4 +1,9 @@
-use crate::{ic::structures::BTreeMap, icu_register_memory, memory::CHILD_INDEX_MEMORY_ID};
+use crate::{
+    Error,
+    ic::structures::BTreeMap,
+    icu_register_memory,
+    memory::{CHILD_INDEX_MEMORY_ID, MemoryError},
+};
 use candid::{CandidType, Principal};
 use derive_more::Deref;
 use serde::{Deserialize, Serialize};
@@ -32,6 +37,10 @@ pub enum ChildIndexError {
 pub struct ChildIndex {}
 
 impl ChildIndex {
+    //
+    // INTERNAL ACCESSORS
+    //
+
     pub fn with<R>(f: impl FnOnce(&BTreeMap<Principal, String>) -> R) -> R {
         CHILD_INDEX.with(|cell| f(&cell.borrow()))
     }
@@ -40,18 +49,39 @@ impl ChildIndex {
         CHILD_INDEX.with(|cell| f(&mut cell.borrow_mut()))
     }
 
+    //
+    // METHODS
+    //
+
     #[must_use]
-    pub fn get_data() -> ChildIndexData {
-        Self::with(|map| ChildIndexData(map.iter_pairs().collect()))
+    pub fn get(pid: &Principal) -> Option<String> {
+        Self::with(|map| map.get(pid))
     }
 
-    pub fn insert_canister(pid: Principal, kind: &str) {
+    pub fn try_get(pid: &Principal) -> Result<String, Error> {
+        if let Some(kind) = Self::get(pid) {
+            Ok(kind)
+        } else {
+            Err(MemoryError::from(ChildIndexError::CanisterNotFound(*pid)))?
+        }
+    }
+
+    #[must_use]
+    pub fn get_by_kind(kind: &str) -> Vec<Principal> {
+        Self::with(|map| {
+            map.iter_pairs()
+                .filter_map(|(p, k)| if k == kind { Some(p) } else { None })
+                .collect()
+        })
+    }
+
+    pub fn insert(pid: Principal, kind: &str) {
         Self::with_mut(|map| {
             map.insert(pid, kind.to_string());
         });
     }
 
-    pub fn remove_canister(pid: &Principal) {
+    pub fn remove(pid: &Principal) {
         Self::with_mut(|map| {
             map.remove(pid);
         });
@@ -63,27 +93,13 @@ impl ChildIndex {
         });
     }
 
-    // get
+    //
+    // EXPORT
+    //
+
     #[must_use]
-    pub fn get(pid: &Principal) -> Option<String> {
-        Self::with(|map| map.get(pid))
-    }
-
-    // try_get
-    pub fn try_get(pid: &Principal) -> Result<String, ChildIndexError> {
-        let canister = Self::get(pid).ok_or(ChildIndexError::CanisterNotFound(*pid))?;
-
-        Ok(canister)
-    }
-
-    // get_by_kind
-    #[must_use]
-    pub fn get_by_kind(kind: &str) -> Vec<Principal> {
-        Self::with(|map| {
-            map.iter_pairs()
-                .filter_map(|(p, k)| if k == kind { Some(p) } else { None })
-                .collect()
-        })
+    pub fn export() -> ChildIndexData {
+        Self::with(|map| ChildIndexData(map.iter_pairs().collect()))
     }
 }