+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - added a way to make ICRC-21 easy
 - made ICRC-10 native
 - added a DelegationCache for other canisters to query an Auth canister
+- made all the CanisterState / Registry errors type 'Error' not the internal Error type
+- rewrote the API for all states and memories to be consistent
 
 ## [0.1.9]
 - added a whole state for session delegations